+++ CHANGELOG.md
@@ -6,6 +6,7 @@ HEAD (unreleased)
 
 Features:
   * Support for FloatOption #156 (Rick Hull)
+  * Support for `limit` config to ArrayOption. (Lee Jarvis)
 
 v4.0.0 (2014-12-27)
 -------------------